Remove unnecessary allocation
The sexp_t_tmpC object allocated at the top of the file is never used.
We copy sexp_gta_oldC into it on line 324, but those values are not
used before sexp_t_tmpC is re-allocated by the call to popEvaluate()
on line 417.

This also created the possibility for invalid memory access, which
would cause R to crash.