Add ability to evaluate compiled objective function
Dirk's RcppDE package has the ability to optimize an objective
function that is written in C/C++, and it often provides nearly an
order of magnitude performance improvement.

Johan Larsson added the ability to pass additional parameters to the
compiled function by passing 2 arguments: 1) parameters to optimize
and 2) an environment containing additional parameters. He credits
this idea to the authors of the Rmalschains package.

Use that pattern to pass a 'NativeSymbol' (i.e. an external pointer)
and an environment to the compiled function. It would be great to
generalize this interface to something more optimizers could use.