Expose compiled objective function API
Other packages need a way to access the objective function pointer.
Add deoptim.h with the function pointer definition.

Update popEvaluate() to loop over the rows of the population matrix
when a function pointer is evaluated. This makes the compiled
objective function easier to write, and is more consistent with other
optimization routines.