issues related to `cluster` argument
For the first one, `cl` has not been defined on L128, for a non-NULL `parVar`, it throws
```r
Error in parallel::clusterExport(cl, ctrl$parVar) : object 'cl' not found
```
although in practice it does not throw an error when `ctrl$parVar` is `NULL`,

For the second one, `cl` is provided outside via `cluster` argument, and it should not be stopped inside. The `stopCluster` operation is for `parallelType=1`, but `parallelType=1` also implies `cluster` is NULL based on the `ifelse` block from L125 to L160. If someone sets `parallelType=1` and a custom `cluster` simultaneously, it would be problematic. So alternatively, another solution might be to disallow to set `parallelType=1` and `cluster` simultaneously.