Update fix to uninitialized value found via valgrind
The previous code did not have an issue, but may be slightly confusing
at first glance. The update makes the intent clearer, and without the
need to check if `i == 0` for each loop iteration.

We declare and initialize `t_bestC = R_PosInf` before the loop. Then
`if (ngta_popC[i] <= t_bestC)` is always TRUE on the first iteration,
which means gt_bestP will always be populated.

Fixes #31.

----------------------------------------------------------------------
For reference, the valgrind issue was in the code below:

    t_bestC = ngta_popC[0];
    for (i = 1; i < i_NP; i++) {
        if (ngta_popC[i] <= t_bestC) {
              t_bestC = ngta_popC[i];
              for (j = 0; j < i_D; j++) {
                  gt_bestP[j]=ngta_popP[i+i_NP*j];
              }
        }
    }

It was possible that `if (ngta_popC[i] <= t_bestC)` would never be
TRUE because the loop started at `i = 1`. Then gt_bestP woudl never be
populated with initial values.
----------------------------------------------------------------------